requirements.txt
requests
numpy
pillow
rasterio
xarray
matplotlib
netcdf4
